+++ comps/AppContent.js
@@ -16,9 +16,9 @@ import Confirmation from './Confirmation';
 import DashboardAccept from './DashboardAccept';
 import DonationPen from '../comps/DonationPen';
 import AccountCreate from './AccountCreate';
-import LoginPage from './LoginPage';
-import SignUp from './SignUp';
-import AcceptingSignUp from './AcceptingSignUp';
+import LoginPage from './login/LoginPage';
+import SignUp from './login/SignUp';
+import AcceptingSignUp from './login/AcceptingSignUp';
 import Donate from './donate';
 import Info from './DInfoMap';
 import DonateContent from './donateContent';
