+++ comps/login/LoginPage.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { View, Text, Image, Button, TextInput, TouchableOpacity, TouchableHighlight, AsyncStorage } from 'react-native';
 import { Actions } from 'react-native-router-flux';
-import Loginstyles from '../styles/LoginStyles';
+import Loginstyles from '../../styles/LoginStyles';
 import axios from 'axios';
 
 
@@ -84,7 +84,7 @@ function LoginPage() {
       <View>
         <Image
           style={Loginstyles.LogoImg}
-          source={require('../assets/img/logo_animation.png')}
+          source={require('../../assets/img/logo_animation.png')}
         />
       </View>
 
@@ -104,7 +104,7 @@ function LoginPage() {
       <View style={Loginstyles.form}>
         <View style={Loginstyles.inputContainer}>
           <Image style={{ width: 20, height: 15, marginLeft: 15, }}
-            source={require('../assets/icon/email.png')} />
+            source={require('../../assets/icon/email.png')} />
           <TextInput style={Loginstyles.inputs}
             placeholder="Email"
             keyboardType="email-address"
@@ -114,7 +114,7 @@ function LoginPage() {
 
         <View style={Loginstyles.inputContainer}>
           <Image style={{ width: 18, height: 25, marginLeft: 15, }}
-            source={require('../assets/icon/password.png')} />
+            source={require('../../assets/icon/password.png')} />
           <TextInput style={Loginstyles.inputs}
             placeholder="Password"
             secureTextEntry={true}
